requirements.txt
PyMuPDF
python-docx
spacy
requests
python-dotenv
pandas
openpyxl